requirements.txt
# Core dependencies
pytesseract>=0.3.10
pdf2image>=1.16.3
python-docx>=0.8.11
Pillow>=9.4.0
Werkzeug>=2.2.3

# Optional dependencies
# Uncomment if you want to use OCRmyPDF as an alternative OCR engine
# ocrmypdf>=14.0.0

# Development dependencies
black>=23.1.0
pylint>=2.16.0